add support for float

+++ calc.py
@@ -5,8 +5,8 @@ def compute(string):
     Example: '5 + 7' -> 12
     """
     values = string.split(' ')
-    num0 = int(values[0])
-    num1 = int(values[2])
+    num0 = float(values[0])
+    num1 = float(values[2])
     operator = values[1]
     if operator == '+':
         return num0 + num1
